ps-bt-streaming/requirements.txt
apache-beam[gcp]
google-cloud-bigtable
